model>serialize+extract/new: add extract method to recover a plain object from a Janus model
Meant for interfacing with other code. Uses the same basic method as
serialize, but leaves in more items.
Also takes a function so that underlying pieces can be loaded before the
extracted object is returned.
Added extract to Varying to automatically trigger on a changed event.
Fixed bug in util with CoffeeScript 1.6.3 compiler