varying/improve: prevent double-fire on reactNow.
reactNow would fire its handler twice initially. this change should
prevent that. all tests pass so I feel reasonable about the change.

the immediate onValue.call in _react along with the logic internal to
onValue handle the initial reaction as expected, so this is extraneous.

one weird thing is that lastRaw doesn’t get set this way, so the value
could theoretically not dedupe as expected the first time. but I suspect
heavily that the lastRaw handling in _bind here doesn’t actually do
anything as deduping is handled further up the chain. but i’ll file a
new ticket to investigate and possibly fix.